cut down log spam resulting from record conversion failures (#38)
we are greatly under reporting these issues and often outright
suppressing.  the issue is we use cloudwatch logging which has
maximum log line limit.  most lines were then split into two log
elements.  the cloudwatch alarm is based on a json logs filter
pattern { $.level="ERROR" } - since the log line is split in two
it can not be parsed as json so is not considered an error.

the log line in question repeated the raw message and error 3 times.
I have moved the raw message to debug and refactored the pattern
where new *Exception or log is applied redundantly to
("text" + cause.getMessage, cause).